requirements.txt
ollama
gitpython
pydantic
pydriller
pygithub
python-dotenv
hypothesis
black
reportlab
textual
customtkinter
pybuilder
google-genai